fix intent for cart:

+++ app/src/app/src/main/java/com/example/ddht/ui/manager/fragment/ManagerProductsFragment.java
@@ -125,7 +125,7 @@ public class ManagerProductsFragment extends Fragment {
                     return;
                 }
                 Intent intent = new Intent(requireContext(), ProductDetailActivity.class);
-                intent.putExtra(ProductDetailActivity.EXTRA_PRODUCT_ID, product.getId());
+                intent.putExtra(ProductDetailActivity.EXTRA_PRODUCT_ID, product.getId().longValue());
                 startActivity(intent);
             }
